% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_alpha}
\alias{add_alpha}
\title{Add alpha to color}
\usage{
add_alpha(col, alpha)
}
\arguments{
\item{col}{a color, e.g., 'red'.}

\item{alpha}{alpha value, range from 0 to 1.}
}
\description{
Add alpha to color
}
\examples{
add_alpha('red', 0.5)
}
