% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippet.R
\name{addPackageSnippets}
\alias{addPackageSnippets}
\title{Export snippets}
\usage{
addPackageSnippets()
}
\value{
boolean invisible(FALSE) if nothing was added, invisible(TRUE) if snipped definitions were added
}
\description{
\code{addPackageSnippets} copies all (missing) snippet definitions
in 'inst/rstudio/Rsnippets.txt' and 'Rmdsnippets.txt' to the RStudios user snippet location.
}
\examples{
\dontrun{addPackageSnippets()}
}
\references{
https://stackoverflow.com/questions/62220596/can-r-packages-add-code-snippets-to-users-snippet-files
}
