% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colname2data}
\alias{colname2data}
\title{Reset \code{colnames} of a data frame}
\usage{
colname2data(data)
}
\arguments{
\item{data}{a \code{data.frame}}
}
\description{
Reset \code{colnames} of a data frame
}
\examples{
colname2data(data.frame(A = 1:3, B = LETTERS[1:3]))
}
