% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_head.R
\docType{data}
\name{.conf}
\alias{.conf}
\alias{.opt}
\title{Mocked global objects to inherit web UI options/settings for plugin developer}
\format{
An object of class \code{list} of length 4.

An object of class \code{list} of length 7.
}
\usage{
.conf

.opt
}
\description{
These provided objects are containers storing default data and information
send from web UI. To help develop plugins, we mock them so we can type
it in R(Studio) and access the available fields.
}
\references{
\itemize{
\item \href{https://hiplot.com.cn/docs/zh/development-guides/#data-json-\%E6\%A0\%BC\%E5\%BC\%8F\%E8\%AF\%B4\%E6\%98\%8E}{Guideline - Data JSON}
\item \href{https://hiplot.com.cn/docs/zh/development-guides/#\%E5\%90\%8E\%E5\%8F\%B0\%E4\%BB\%BB\%E5\%8A\%A1\%E4\%BB\%A3\%E7\%A0\%81\%E8\%AF\%B4\%E6\%98\%8E}{后台任务代码说明}
}
}
\keyword{datasets}
