% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R, R/import.R
\name{export}
\alias{export}
\alias{export_single}
\alias{export_pptx}
\alias{export_plotly}
\alias{export_directory}
\alias{export_htmlwidget}
\alias{html2pdf}
\alias{pdf2image}
\alias{pdfs2image}
\alias{merge_pdfs}
\alias{split_pdfs}
\alias{import_images_to_pdf}
\alias{read_data}
\alias{parse_file_link}
\title{Functions and utilities to export hiplot output}
\usage{
export_single(x)

export_pptx(x, outfn, width, height)

export_plotly(x, outfn)

export_directory(
  outdir = "output",
  pdf_tb_subdir = c("output", "output/results"),
  recursive = c(FALSE, TRUE),
  sheetname_replace = "_BayesianNMF|legacy_fitting_|_fitting|.csv"
)

export_htmlwidget(x)

html2pdf(inhtml, outpdf)

pdf2image(pdf, i = "")

pdfs2image(pdfs)

merge_pdfs(pdfs, outpdf)

split_pdfs(pdfs, outdir)

import_images_to_pdf(pdfs)

read_data(path, format = c("csv", "excel", "rds", "rdata"), header = TRUE, ...)

parse_file_link(link)
}
\arguments{
\item{x}{a plot object to export.}

\item{outfn}{output file path.}

\item{width}{width of output pptx.}

\item{height}{height of output pptx.}

\item{outdir}{output directory.}

\item{pdf_tb_subdir}{directory path storing pdf (figure) files.}

\item{recursive}{same path as pdf_tb_subdir to specify if run with recursive mode.}

\item{sheetname_replace}{all .csv files are re-exported into a excel file,
if necessary, use this option remove some string as sheet names.}

\item{inhtml}{input html file.}

\item{outpdf}{output pdf file.}

\item{pdf}{input pdf file.}

\item{i}{a postfix added to the output file name.}

\item{pdfs}{input pdf files.}

\item{path}{the file path to data.}

\item{format}{import data format.}

\item{header}{if \code{TRUE}, tread input having a header.}

\item{...}{other parameters passing to \link[data.table:fread]{data.table::fread} when use
this function to import inside.}

\item{link}{the file link}
}
\description{
Functions and utilities to export hiplot output
}
\section{Functions}{
\itemize{
\item \code{export_single}: Export a single ggplot object

\item \code{export_pptx}: Export a single ggplot object to PPTx

\item \code{export_plotly}: Export a single ggplot object to plotly

\item \code{export_directory}: Export hiplot plugin output directory to a zipped file

\item \code{export_htmlwidget}: Export a single ggplot object to html file

\item \code{html2pdf}: Convert html to pdf

\item \code{pdf2image}: Convert pdf to images

\item \code{pdfs2image}: Convert multiple pdfs to images

\item \code{merge_pdfs}: Merge multiple pdf files into one

\item \code{split_pdfs}: Split multiple page pdfs to a directory

\item \code{import_images_to_pdf}: Export non-pdf images as pdfs

\item \code{read_data}: Read common format data into data frame

\item \code{parse_file_link}: Parse the true file path from link
}}

\examples{
# parse_file_link("hiplot.org/path=xxx") TODO
}
