% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\docType{data}
\name{ggplot2}
\alias{ggplot2}
\alias{ggplot2_themes}
\alias{ggplot2_themes2}
\alias{choose_ggplot_theme}
\alias{choose_ggplot_theme2}
\title{Prestored settings and functions related to ggplot2}
\format{
An object of class \code{character} of length 50.

An object of class \code{list} of length 6.
}
\usage{
ggplot2_themes

ggplot2_themes2

choose_ggplot_theme(p, theme = "theme_bw")

choose_ggplot_theme2(theme = "theme_bw")
}
\arguments{
\item{p}{a ggplot object.}

\item{theme}{a string specifying theme.}
}
\description{
Prestored settings and functions related to ggplot2
}
\section{Functions}{
\itemize{
\item \code{ggplot2_themes}: a vector of ggplot2 themes

\item \code{ggplot2_themes2}: a list of ggplot2 themes

\item \code{choose_ggplot_theme}: Set ggplot2 theme for a ggplot object

\item \code{choose_ggplot_theme2}: Return ggplot2 theme based on string
}}

\keyword{datasets}
