% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_head.R
\name{globs}
\alias{globs}
\alias{globs_list}
\alias{globs_get}
\alias{globs_set}
\title{Functions to manage the global setting objects}
\usage{
globs_list()

globs_get(x)

globs_set(x, name = NULL)
}
\arguments{
\item{x}{the name of the global object, the most common used is "conf".}

\item{name}{if x is not a valid global object name, then specify it with \code{name}.
This is not recommended.}
}
\description{
Functions to manage the global setting objects
}
\section{Functions}{
\itemize{
\item \code{globs_list()}: list available global objects

\item \code{globs_get()}: get the data from specified global objects

\item \code{globs_set()}: reassign the specified data (with name) to available global object

}}
\examples{
\dontrun{
conf <- globs_get("conf")
# Modify conf
# xxx
# Then reassign
globs_set("conf")
}
}
