% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiplotlib.R
\name{hiplotlib}
\alias{hiplotlib}
\title{hiplotlib package}
\description{
Infrastructure functions for Hiplot plot system.
}
\seealso{
Useful links:

\url{https://github.com/hiplot/hiplotlib}

Report bugs at \url{https://github.com/hiplot/hiplotlib/issues}
}
