% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_same_string_position}
\alias{list_same_string_position}
\title{Get the same characters in two strings}
\usage{
list_same_string_position(
  a,
  b,
  exclude = c("-", "?"),
  ignore_case = TRUE,
  show_excluded = FALSE,
  only_position = TRUE
)
}
\arguments{
\item{a, b}{a string.}

\item{exclude}{symbol to exclude from comparison.}

\item{ignore_case}{if \code{TRUE}, ignore case.}

\item{show_excluded}{if \code{TRUE}, show excluded.}

\item{only_position}{if \code{TRUE}, only position vector returned.}
}
\description{
Get the same characters in two strings
}
\examples{
list_same_string_position("abcde", "fbcde")
list_same_string_position("abcde", "fbcde", only_position = FALSE)
}
