% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiplot.R
\name{run_hiplot}
\alias{run_hiplot}
\title{Run hiplot}
\usage{
run_hiplot(opt = globs_get("opt"))
}
\description{
Run hiplot
}
\examples{
\dontrun{
basedir = system.file("extdata", "ezcox", package = "hiplotlib")
opt = list(inputFile = file.path(basedir, "data.txt"),
           confFile = file.path(basedir, "data.json"),
           outputFilePrefix = file.path(basedir, "result/test"),
           tool = "ezcox",
           module = "basic",
           simple = FALSE,
           enableExample = TRUE,
           help = FALSE)
dir.create(dirname(opt$outputFilePrefix))
dir.create(file.path(dirname(opt$outputFilePrefix), "log"))
options(hiplotlib.script_dir = dirname(basedir))
run_hiplot()
}
}
