% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_factors}
\alias{set_factors}
\title{Set a vector to a factor}
\usage{
set_factors(x, rev = FALSE, sort = FALSE)
}
\arguments{
\item{x}{a vector.}

\item{rev}{if \code{TRUE}, reverse the factor level.}

\item{sort}{if \code{TRUE}, sort the vector before setting factor.}
}
\description{
Set a vector to a factor
}
\examples{
set_factors(c("B", "B", "A", "C"))
set_factors(c("B", "B", "A", "C"), rev = TRUE)
set_factors(c("B", "B", "A", "C"), sort = TRUE)
}
