% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_head.R
\name{set_global_confs}
\alias{set_global_confs}
\title{Check and update global setting before starting plugin}
\usage{
set_global_confs()
}
\description{
Check and update global setting before starting plugin
}
