% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{system_safe}
\alias{system_safe}
\title{Run system commands}
\usage{
system_safe(cmd, print_cmd = TRUE)
}
\arguments{
\item{cmd}{the command to run.}

\item{print_cmd}{if \code{TRUE}, also print the command.}
}
\description{
Run system commands
}
