% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task}
\alias{task}
\alias{update_task_status}
\alias{new_task_step}
\alias{run_task_step}
\title{Functions and utilities related to hiplot task}
\usage{
update_task_status(conf)

new_task_step(id, label, func)

run_task_step(id, func)
}
\arguments{
\item{conf}{Web UI conf.}

\item{id}{task id}

\item{label}{task label}

\item{func}{a function to run}
}
\description{
Functions and utilities related to hiplot task
}
\section{Functions}{
\itemize{
\item \code{update_task_status()}: Update task status

\item \code{new_task_step()}: Create a task to run

\item \code{run_task_step()}: Run a step of a task #TODO What's the difference from above?

}}
