% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform}
\alias{transform}
\alias{transform_val}
\title{Transform value}
\usage{
transform_val(func_str, val)
}
\arguments{
\item{func_str}{a string represents a function.}

\item{val}{value passing to function}
}
\description{
Transform value
}
