% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unlist_and_covert}
\alias{unlist_and_covert}
\title{Unlist and convert "NA" to NA}
\usage{
unlist_and_covert(x, recursive = FALSE)
}
\arguments{
\item{x}{data.}

\item{recursive}{if \code{TRUE}, unlist recursively.}
}
\description{
Unlist and convert "NA" to NA
}
\examples{
unlist_and_covert(NULL)
unlist_and_covert(list())
unlist_and_covert(list("A", "B"))
unlist_and_covert(list("NA", "B"))
unlist_and_covert(list(list(A = 1), "B"))
}
