% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize a string}
\usage{
capitalize(string, locale = "en")
}
\arguments{
\item{string}{String to modify}

\item{locale}{Locale to use for translations. Defaults to "en" (English)
to ensure consistent default ordering across platforms.}
}
\description{
Capitalize a string
}
\examples{
capitalize("aBcDeF")
}
