% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{system_safe}
\alias{system_safe}
\title{Run system commands}
\usage{
system_safe(cmd, print_cmd = TRUE, conda_env = NULL)
}
\arguments{
\item{cmd}{the command to run.}

\item{print_cmd}{if \code{TRUE}, also print the command.}

\item{conda_env}{set conda env.}
}
\description{
Run system commands
}
